+++ app/client/src/pages/instructor/ManageCourse.jsx
@@ -5,45 +5,24 @@ import Navbar from "../../components/Navbar";
 import CreateCourse from "../../components/instructor/modals/CreateCourse";
 import { getUserCourseList } from "../../api/course/courses";
 import { useAuth } from "../../AuthContext";
-import { useNavigate } from "react-router-dom";
 import { DataGrid } from "@mui/x-data-grid";
 import { tableColumns, tableRows } from "../../utilis/ManageCourseTable";
 
 const defaultTheme = createTheme();
 
 const ManageCourse = () => {
-  const [page, setPage] = useState(0);
-  const [rowsPerPage, setRowsPerPage] = useState(10);
   const [open, setOpen] = useState(false);
   const [successMsg, setSuccessMsg] = useState("");
   const [alertOpen, setAlertOpen] = useState(false);
   const [courses, setCourses] = useState([]);
-  const handleOpen = () => setOpen(true);
-  const handleClose = () => setOpen(false);
   const { userId } = useAuth();
-  const navigator = useNavigate();
+
 
   const columns = tableColumns;
   const rows = tableRows(courses);
 
-  const handleChangePage = (event, newPage) => {
-    setPage(newPage);
-  };
-
-  const handleChangeRowsPerPage = (event) => {
-    setRowsPerPage(+event.target.value);
-    setPage(0);
-  };
-
-  const handleEnter = (course) => {
-    navigator(`/c/${course.crn}`, {
-      state: {
-        courseCRN: course.crn,
-        courseYear: course.year,
-        courseSemester: course.semester,
-      },
-    });
-  };
+  const handleOpen = () => setOpen(true);
+  const handleClose = () => setOpen(false);
 
   useEffect(() => {
     const getCourseList = async () => {
@@ -61,7 +40,7 @@ const ManageCourse = () => {
   return (
     <ThemeProvider theme={defaultTheme}>
       <CssBaseline />
-      <Navbar name="Manage Courses" redirect={true} />
+      <Navbar name="Course List" redirect={true} />
       <Container>
         <Grid
           container
