+++ app/client/src/utilis/ManageCourseTable.js
@@ -1,3 +1,31 @@
+import React from "react";
+import { Button } from "@mui/material";
+import { useNavigate } from "react-router-dom";
+
+
+
+
+const ViewButton = ({ row }) => {
+    const navigator = useNavigate();
+    const handleEnter = (course) => {
+        navigator(`/c/${course.crn}`, {
+            state: {
+                courseCRN: course.crn,
+                courseYear: course.year,
+                courseSemester: course.semester,
+            },
+        });
+    };
+
+    return (<Button onClick={() => handleEnter(row)}
+        style={{ backgroundColor: "#2eb24d" }}
+        variant="contained">
+        View
+    </Button>
+    );
+
+}
+
 export const tableColumns = [
     { field: "name", headerName: "Course Name", width: 170 },
     { field: "code", headerName: "Code", width: 100 },
@@ -27,9 +55,12 @@ export const tableColumns = [
         width: 170,
     },
     {
-        field: "enter",
-        headerName: "Enter",
+        field: "view",
+        headerName: "View",
         width: 170,
+        sortable: false,
+        renderCell: ({ row }) => <ViewButton row={row} />,
+
     },
 ]
